requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
pydantic==2.9.2
python-decouple==3.8
httpx==0.27.0
tenacity==9.0.0
pytest==8.1.1
pytest-asyncio==0.23.6
